+++ Debug/Types/CharacterSheet.lua
@@ -1,0 +1,16 @@
+---@alias LCS_SlotSide "LEFT" | "RIGHT"
+
+---@class LCS_Slot
+---@field id number
+---@field side LCS_SlotSide
+---@field name string
+---@field canEnchant boolean
+
+---@class LCS_SlotOverlay : Frame
+---@field Tint Texture
+---@field Level FontString
+---@field MaxLevel FontString
+---@field Enchant FontString
+---@field Sockets Button[]
+
+---@class LCS_Events : Frame