+++ Debug/Types/Config.lua
@@ -1,0 +1,24 @@
+---@class LCS_ConfigColors
+---@field enchants ColorTable
+---@field enchantsMissing ColorTable
+---@field texture ColorTable
+---@field levels ColorTable
+---@field maxLevels ColorTable
+---@field maxLevelsUpgraded ColorTable
+
+---@class LCS_ConfigFonts
+---@field font string
+---@field size number
+---@field outline string
+
+---@class LCS_ConfigShow
+---@field levels boolean
+---@field maxLevels boolean
+---@field enchants boolean
+---@field enchantsMissing boolean
+---@field gems boolean
+
+---@class LCS_Config
+---@field colors LCS_ConfigColors
+---@field fonts LCS_ConfigFonts
+---@field show LCS_ConfigShow