+++ LiqCharacterSheet.lua
@@ -1,5 +1,3 @@
----@type string
-local addonName = select(1, ...)
 ---@class LCS_Addon
 local addon = select(2, ...)
 
@@ -13,13 +11,11 @@ local defaultColorEnchantMissing = {r = 1, g = 0, b = 0, a = 1}
 local defaultColorTexture = {r = 0, g = 0, b = 0, a = 0.33}
 local defaultColorLevel = {r = 1, g = 1, b = 1, a = 1}
 local defaultColorMaxLevel = {r = 1, g = 1, b = 1, a = 1}
-local defaultColormaxLevelsUpgraded = {r = 0, g = 1, b = 0, a = 1}
-local LCSFrameName = "LCSFrame"
+local defaultColorMaxLevelsUpgraded = {r = 0, g = 1, b = 0, a = 1}
+local slotOverlayFrameName = "LCSFrame"
+local maxUpgradeLevel = 289
 
---@debug@
-_G[addonName] = addon
---@end-debug@
-
+---@type LCS_Config
 local config = {
   colors = {
     enchants = defaultColorEnchant,
@@ -27,7 +23,7 @@ local config = {
     texture = defaultColorTexture,
     levels = defaultColorLevel,
     maxLevels = defaultColorMaxLevel,
-    maxLevelsUpgraded = defaultColormaxLevelsUpgraded,
+    maxLevelsUpgraded = defaultColorMaxLevelsUpgraded,
   },
   fonts = {
     font = defaultFont,
@@ -43,6 +39,7 @@ local config = {
   },
 }
 
+---@type table<number, LCS_Slot>
 local Slots = {
   [1] = {id = 1, side = "LEFT", name = "Head", canEnchant = true},
   [2] = {id = 2, side = "LEFT", name = "Neck", canEnchant = false},
@@ -61,13 +58,13 @@ local Slots = {
   [15] = {id = 15, side = "LEFT", name = "Back", canEnchant = false},
   [16] = {id = 16, side = "RIGHT", name = "MainHand", canEnchant = true},
   [17] = {id = 17, side = "LEFT", name = "SecondaryHand", canEnchant = true},
-  --    [18] = {id = 18, side = "LEFT", name = "Ranged", canEnchant = false},
-  --    [19] = {id = 19, side = "LEFT", name = "Tabard", canEnchant = false}
+  -- [18] = {id = 18, side = "LEFT", name = "Ranged", canEnchant = false},
+  -- [19] = {id = 19, side = "LEFT", name = "Tabard", canEnchant = false},
 }
 
--- Credit: https://www.raidbots.com/static/data/live/bonuses.json
--- Season 17 / Midnight S1
-local maxUpgaradeLevel = 289
+--- Credit: https://www.raidbots.com/static/data/live/bonuses.json
+--- Season 17 / Midnight S1
+---@type table<number, number[]>
 local maxUpgradeLevels = {
   [237] = {12769, 12770, 12771, 12772, 12773, 12774},                   -- Adventurer
   [250] = {12777, 12778, 12779, 12780, 12781, 12782},                   -- Veteran
@@ -77,10 +74,12 @@ local maxUpgradeLevels = {
   [285] = {9401, 9402, 9403, 9404, 9405, 9623, 9624, 9625, 9626, 9627}, -- Crafted Qualities
 }
 
-local GetMaxUpgradeLevel = function(bonusId)
+---@param bonusId number
+---@return number|nil
+local function GetMaxUpgradeLevel(bonusId)
   for level, ids in pairs(maxUpgradeLevels) do
-    for i, id in pairs(ids) do
-      if id == bonusId then
+    for index = 1, #ids do
+      if ids[index] == bonusId then
         return level
       end
     end
@@ -88,99 +87,100 @@ local GetMaxUpgradeLevel = function(bonusId)
   return nil
 end
 
-local UpdateSlot = function(unitId, slotId)
-  if unitId == nil or slotId == nil then return end
-
-  local slot = Slots[slotId]
-  if slot == nil then return end
-
-  -- Blizzard inventory slot frame
-  local CharacterSlotFrame = _G[(unitId == "player" and "Character" or "Inspect") .. slot.name .. "Slot"]
-  if CharacterSlotFrame == nil then
-    return nil
+---@param unitId string
+---@param slotId number
+local function UpdateSlot(unitId, slotId)
+  if unitId == nil or slotId == nil then
+    return
   end
 
-  -- Get or create new frames
-  local LCSFrame = CharacterSlotFrame[LCSFrameName]
-  if LCSFrame == nil then
+  local slot = Slots[slotId]
+  if slot == nil then
+    return
+  end
+
+  ---@type Button|nil
+  local characterSlotFrame = _G[(unitId == "player" and "Character" or "Inspect") .. slot.name .. "Slot"]
+  if characterSlotFrame == nil then
+    return
+  end
+
+  ---@type LCS_SlotOverlay|nil
+  local slotOverlay = characterSlotFrame[slotOverlayFrameName]
+  if slotOverlay == nil then
     local relativePoint = slot.side == "LEFT" and "RIGHT" or "LEFT"
     local offsetX = slot.side == "LEFT" and 9 or -10
     local offsetEnchantY = (slot.id == 16 or slot.id == 17) and -12 or 8
 
-    CharacterSlotFrame[LCSFrameName] = CreateFrame("Frame", CharacterSlotFrame:GetName() .. LCSFrameName, CharacterSlotFrame)
-    LCSFrame = CharacterSlotFrame[LCSFrameName]
-    LCSFrame:SetPoint("CENTER")
-    LCSFrame:SetAllPoints(CharacterSlotFrame)
+    characterSlotFrame[slotOverlayFrameName] = CreateFrame("Frame", characterSlotFrame:GetName() .. slotOverlayFrameName, characterSlotFrame)
+    slotOverlay = characterSlotFrame[slotOverlayFrameName]
 
-    -- Overlay tint
-    LCSFrame.Tint = LCSFrame:CreateTexture(nil, "BACKGROUND")
-    LCSFrame.Tint:SetTexture("Interface\\TutorialFrame\\TutorialFrameBackground")
-    LCSFrame.Tint:SetAllPoints(LCSFrame)
+    slotOverlay:SetPoint("CENTER")
+    slotOverlay:SetAllPoints(characterSlotFrame)
 
-    -- Level string
-    LCSFrame.Level = LCSFrame:CreateFontString(LCSFrame:GetName() .. "Level", "OVERLAY", "GameTooltipText")
-    LCSFrame.Level:SetPoint("CENTER", LCSFrame, "CENTER", 0, 0)
-    local font, fontSize = LCSFrame.Level:GetFont()
+    slotOverlay.Tint = slotOverlay:CreateTexture(nil, "BACKGROUND")
+    slotOverlay.Tint:SetTexture("Interface\\TutorialFrame\\TutorialFrameBackground")
+    slotOverlay.Tint:SetAllPoints(slotOverlay)
+
+    slotOverlay.Level = slotOverlay:CreateFontString(slotOverlay:GetName() .. "Level", "OVERLAY", "GameTooltipText")
+    slotOverlay.Level:SetPoint("CENTER", slotOverlay, "CENTER", 0, 0)
+    local font, fontSize = slotOverlay.Level:GetFont()
     if font then
-      LCSFrame.Level:SetFont(font, fontSize + 1, "OUTLINE")
+      slotOverlay.Level:SetFont(font, fontSize + 1, "OUTLINE")
     end
 
-    -- Max level string
-    LCSFrame.MaxLevel = LCSFrame:CreateFontString(LCSFrame:GetName() .. "MaxLevel", "OVERLAY", "GameTooltipText")
-    LCSFrame.MaxLevel:SetPoint("CENTER", LCSFrame, "CENTER", 0, -8)
+    slotOverlay.MaxLevel = slotOverlay:CreateFontString(slotOverlay:GetName() .. "MaxLevel", "OVERLAY", "GameTooltipText")
+    slotOverlay.MaxLevel:SetPoint("CENTER", slotOverlay, "CENTER", 0, -8)
     if font then
-      LCSFrame.MaxLevel:SetFont(font, fontSize - 3, "OUTLINE")
+      slotOverlay.MaxLevel:SetFont(font, fontSize - 3, "OUTLINE")
     end
 
-    -- Enchant string
-    LCSFrame.Enchant = LCSFrame:CreateFontString(LCSFrame:GetName() .. "Enchant", "OVERLAY", "GameTooltipText")
-    LCSFrame.Enchant:SetPoint(slot.side, LCSFrame, relativePoint, offsetX, offsetEnchantY)
-    LCSFrame.Enchant:SetWidth(80)
-    LCSFrame.Enchant:SetWordWrap(false)
+    slotOverlay.Enchant = slotOverlay:CreateFontString(slotOverlay:GetName() .. "Enchant", "OVERLAY", "GameTooltipText")
+    slotOverlay.Enchant:SetPoint(slot.side, slotOverlay, relativePoint, offsetX, offsetEnchantY)
+    slotOverlay.Enchant:SetWidth(80)
+    slotOverlay.Enchant:SetWordWrap(false)
     if font then
-      LCSFrame.Enchant:SetFont(font, fontSize - 3, "OUTLINE")
+      slotOverlay.Enchant:SetFont(font, fontSize - 3, "OUTLINE")
     end
 
-    -- Sockets
-    LCSFrame.Sockets = {}
-    for i = 1, 3 do
-      if LCSFrame.Sockets[i] == nil then
-        LCSFrame.Sockets[i] = CreateFrame("Button", LCSFrame:GetName() .. "Socket" .. i, LCSFrame, "UIPanelButtonTemplate")
-        LCSFrame.Sockets[i]:SetSize(14, 14)
-        local socketOffsetX = offsetX - 3 - (15 * (i - 1))
+    slotOverlay.Sockets = {}
+    for socketIndex = 1, 3 do
+      if slotOverlay.Sockets[socketIndex] == nil then
+        slotOverlay.Sockets[socketIndex] = CreateFrame("Button", slotOverlay:GetName() .. "Socket" .. socketIndex, slotOverlay, "UIPanelButtonTemplate")
+        slotOverlay.Sockets[socketIndex]:SetSize(14, 14)
+        local socketOffsetX = offsetX - 3 - (15 * (socketIndex - 1))
         if slot.side == "LEFT" then
-          socketOffsetX = offsetX + 3 + (15 * (i - 1))
+          socketOffsetX = offsetX + 3 + (15 * (socketIndex - 1))
         end
-        LCSFrame.Sockets[i]:SetPoint(slot.side, LCSFrame:GetName(), relativePoint, socketOffsetX, 0)
+        slotOverlay.Sockets[socketIndex]:SetPoint(slot.side, slotOverlay:GetName(), relativePoint, socketOffsetX, 0)
       end
     end
   end
 
   local itemId = GetInventoryItemID(unitId, slotId)
   if itemId == nil then
-    LCSFrame:Hide()
+    slotOverlay:Hide()
     return
   end
 
   local itemLink = GetInventoryItemLink(unitId, slotId)
   if itemLink == nil or itemLink == "" then
-    LCSFrame:Hide()
+    slotOverlay:Hide()
     return
   end
 
   local itemPayload = string.match(itemLink, "item:([%-?%d:]+)")
   if itemPayload == nil then
-    LCSFrame:Hide()
+    slotOverlay:Hide()
     return
   end
 
-  -- local itemRarityColor = string.sub(itemLink, 3, 10)
   local itemPayloadSplit = {strsplit(":", itemPayload)}
   local itemEnchant = nil
   local itemEnchantAtlas = ""
   local itemSocketCount = 0
   local itemSockets = {}
-  local MaxLevel = nil
+  local maxLevel = nil
   local enchantPattern = ENCHANTED_TOOLTIP_LINE:gsub("%%s", "(.*)")
   local enchantAtlasPattern = "(.*)|A:(.*):20:20|a"
   local enchantText = ""
@@ -189,7 +189,7 @@ local UpdateSlot = function(unitId, slotId)
   local colorTexture = defaultColorTexture
   local colorLevel = defaultColorLevel
   local colorMaxLevel = defaultColorMaxLevel
-  local colorMaxLevelUpgraded = defaultColormaxLevelsUpgraded
+  local colorMaxLevelUpgraded = defaultColorMaxLevelsUpgraded
   local font = defaultFont
   local fontSize = defaultFontsize
   local fontOutline = defaultFontOutline
@@ -226,38 +226,39 @@ local UpdateSlot = function(unitId, slotId)
   end
 
   local itemLevel = nil
-  local tooltipData = C_TooltipInfo.GetHyperlink(itemLink)
-  if tooltipData ~= nil and tooltipData.lines ~= nil then
-    for _, lineData in ipairs(tooltipData.lines) do
+  local hyperlinkTooltipData = C_TooltipInfo.GetHyperlink(itemLink)
+  if hyperlinkTooltipData ~= nil and hyperlinkTooltipData.lines ~= nil then
+    for lineIndex = 1, #hyperlinkTooltipData.lines do
+      local lineData = hyperlinkTooltipData.lines[lineIndex]
       if lineData.type == Enum.TooltipDataLineType.ItemLevel then
         itemLevel = lineData.itemLevel
       end
     end
   end
 
-  -- local itemLevel = C_Item.GetDetailedItemLevelInfo(itemLink)
   local _, _, _, _, _, _, _, _, itemEquipLoc = C_Item.GetItemInfo(itemLink)
   if itemLevel == nil or config.show.levels == false or config.show.levels == nil then
-    LCSFrame.Level:Hide()
-    LCSFrame.MaxLevel:Hide()
-    LCSFrame.Tint:Hide()
+    slotOverlay.Level:Hide()
+    slotOverlay.MaxLevel:Hide()
+    slotOverlay.Tint:Hide()
   else
-    LCSFrame.Level:SetText(tostring(itemLevel))
-    LCSFrame.Level:Show()
-    LCSFrame.Tint:Show()
+    slotOverlay.Level:SetText(tostring(itemLevel))
+    slotOverlay.Level:Show()
+    slotOverlay.Tint:Show()
     if config.show.maxLevels == true then
-      LCSFrame.MaxLevel:Show()
+      slotOverlay.MaxLevel:Show()
     else
-      LCSFrame.MaxLevel:Hide()
+      slotOverlay.MaxLevel:Hide()
     end
   end
 
   local upgradePattern = ITEM_UPGRADE_TOOLTIP_FORMAT_STRING
   upgradePattern = upgradePattern:gsub("%%d", "%%s")
   upgradePattern = upgradePattern:format("(.+)", "(%d+)", "(%d+)")
-  local tooltipData = C_TooltipInfo.GetInventoryItem(unitId, slotId)
-  if tooltipData ~= nil then
-    for i, line in pairs(tooltipData.lines) do
+  local inventoryTooltipData = C_TooltipInfo.GetInventoryItem(unitId, slotId)
+  if inventoryTooltipData ~= nil then
+    for lineIndex = 1, #inventoryTooltipData.lines do
+      local line = inventoryTooltipData.lines[lineIndex]
       local text = line.leftText
       local enchantString = string.match(text, enchantPattern)
       if enchantString ~= nil then
@@ -281,8 +282,8 @@ local UpdateSlot = function(unitId, slotId)
       local match = line.leftText:find(upgradePattern)
       if match then
         if line.leftColor:GenerateHexColor() == DISABLED_FONT_COLOR:GenerateHexColor() then
-          local r, g, b, a = DISABLED_FONT_COLOR:GetRGBA()
-          colorLevel = {r = r, g = g, b = b, a = a}
+          local red, green, blue, alpha = DISABLED_FONT_COLOR:GetRGBA()
+          colorLevel = {r = red, g = green, b = blue, a = alpha}
         end
       end
     end
@@ -290,42 +291,40 @@ local UpdateSlot = function(unitId, slotId)
 
   local numBonuses = tonumber(itemPayloadSplit[13])
   if numBonuses ~= nil and numBonuses > 0 then
-    for i = 14, 13 + numBonuses do
-      local bonusId = tonumber(itemPayloadSplit[i])
+    for bonusIndex = 14, 13 + numBonuses do
+      local bonusId = tonumber(itemPayloadSplit[bonusIndex])
       if bonusId ~= nil then
         local maxLevelUpgrade = GetMaxUpgradeLevel(bonusId)
         if maxLevelUpgrade ~= nil then
-          MaxLevel = maxLevelUpgrade
+          maxLevel = maxLevelUpgrade
         end
       end
     end
   end
 
-  ---TWW S3: Reshii Wraps
   if itemId == 235499 then
-    MaxLevel = 170
+    maxLevel = 170
   end
 
-  ---TWW S2: D.I.S.C Belt
   if itemId == 245966 or itemId == 245964 or itemId == 245965 or itemId == 242664 then
-    MaxLevel = 141
+    maxLevel = 141
   end
 
-  if MaxLevel == nil then
+  if maxLevel == nil then
     if config.colors.levels ~= nil then
-      local r, g, b, a = DISABLED_FONT_COLOR:GetRGBA()
-      colorLevel = {r = r, g = g, b = b, a = a}
+      local red, green, blue, alpha = DISABLED_FONT_COLOR:GetRGBA()
+      colorLevel = {r = red, g = green, b = blue, a = alpha}
     end
-    LCSFrame.Level:SetPoint("CENTER", LCSFrame, "CENTER", 0, 0)
-    LCSFrame.MaxLevel:Hide()
+    slotOverlay.Level:SetPoint("CENTER", slotOverlay, "CENTER", 0, 0)
+    slotOverlay.MaxLevel:Hide()
   else
-    LCSFrame.MaxLevel:SetText(tostring(MaxLevel))
+    slotOverlay.MaxLevel:SetText(tostring(maxLevel))
     if config.show.levels == true and config.show.maxLevels == true then
-      LCSFrame.Level:SetPoint("CENTER", LCSFrame, "CENTER", 0, 4)
-      LCSFrame.MaxLevel:Show()
+      slotOverlay.Level:SetPoint("CENTER", slotOverlay, "CENTER", 0, 4)
+      slotOverlay.MaxLevel:Show()
     else
-      LCSFrame.Level:SetPoint("CENTER", LCSFrame, "CENTER", 0, 0)
-      LCSFrame.MaxLevel:Hide()
+      slotOverlay.Level:SetPoint("CENTER", slotOverlay, "CENTER", 0, 0)
+      slotOverlay.MaxLevel:Hide()
     end
   end
 
@@ -334,12 +333,12 @@ local UpdateSlot = function(unitId, slotId)
       enchantText = "No enchant"
       colorEnchant = colorEnchantMissing
       if config.show.enchantsMissing == true and (itemEquipLoc ~= "INVTYPE_HOLDABLE" and itemEquipLoc ~= "INVTYPE_SHIELD") then
-        LCSFrame.Enchant:Show()
+        slotOverlay.Enchant:Show()
       else
-        LCSFrame.Enchant:Hide()
+        slotOverlay.Enchant:Hide()
       end
     else
-      LCSFrame.Enchant:Hide()
+      slotOverlay.Enchant:Hide()
     end
   else
     itemEnchant = itemEnchant:gsub("+", "")
@@ -349,36 +348,34 @@ local UpdateSlot = function(unitId, slotId)
       enchantText = itemEnchant
     end
     if config.show.enchants == true then
-      LCSFrame.Enchant:Show()
+      slotOverlay.Enchant:Show()
     else
-      LCSFrame.Enchant:Hide()
+      slotOverlay.Enchant:Hide()
     end
   end
 
-  LCSFrame.Enchant:SetText(enchantText)
-  LCSFrame.Enchant:SetTextColor(colorEnchant.r, colorEnchant.g, colorEnchant.b, colorEnchant.a)
-  LCSFrame.Enchant:SetJustifyH(slot.side == "RIGHT" and "RIGHT" or "LEFT")
+  slotOverlay.Enchant:SetText(enchantText)
+  slotOverlay.Enchant:SetTextColor(colorEnchant.r, colorEnchant.g, colorEnchant.b, colorEnchant.a)
+  slotOverlay.Enchant:SetJustifyH(slot.side == "RIGHT" and "RIGHT" or "LEFT")
 
-  -- Weapon enchants
   if slot.id ~= 16 and slot.id ~= 17 then
-    local point, relativeTo, relativePoint, offset_x = LCSFrame.Enchant:GetPoint()
+    local point, relativeTo, relativePoint, offsetX = slotOverlay.Enchant:GetPoint()
     if itemSocketCount > 0 or (slot.id == 9 or slot.id == 14) then
-      LCSFrame.Enchant:SetPoint(point, relativeTo, relativePoint, offset_x, 8)
+      slotOverlay.Enchant:SetPoint(point, relativeTo, relativePoint, offsetX, 8)
     else
-      LCSFrame.Enchant:SetPoint(point, relativeTo, relativePoint, offset_x, 0)
+      slotOverlay.Enchant:SetPoint(point, relativeTo, relativePoint, offsetX, 0)
     end
   end
 
-  -- Sockets
-  for i = 1, 3 do
-    local _, gemLink = C_Item.GetItemGem(itemLink, i)
-    local socketFrame = LCSFrame.Sockets[i]
-    local point, relativeTo, relativePoint, offset_x = socketFrame:GetPoint()
+  for socketIndex = 1, 3 do
+    local _, gemLink = C_Item.GetItemGem(itemLink, socketIndex)
+    local socketFrame = slotOverlay.Sockets[socketIndex]
+    local point, relativeTo, relativePoint, offsetX = socketFrame:GetPoint()
 
     if gemLink == nil then
-      if i <= itemSocketCount then
-        if itemSockets[i] ~= nil and config.show.gems == true then
-          socketFrame:SetNormalTexture(itemSockets[i])
+      if socketIndex <= itemSocketCount then
+        if itemSockets[socketIndex] ~= nil and config.show.gems == true then
+          socketFrame:SetNormalTexture(itemSockets[socketIndex])
           socketFrame:Show()
         else
           socketFrame:Hide()
@@ -395,8 +392,8 @@ local UpdateSlot = function(unitId, slotId)
       socketFrame:SetScript("OnLeave", function()
         GameTooltip:Hide()
       end)
-      if itemSockets[i] ~= nil and config.show.gems == true then
-        socketFrame:SetNormalTexture(itemSockets[i])
+      if itemSockets[socketIndex] ~= nil and config.show.gems == true then
+        socketFrame:SetNormalTexture(itemSockets[socketIndex])
         socketFrame:Show()
       else
         socketFrame:Hide()
@@ -404,67 +401,85 @@ local UpdateSlot = function(unitId, slotId)
     end
 
     if enchantText ~= "" or (slot.id == 9 or slot.id == 14) then
-      socketFrame:SetPoint(point, relativeTo, relativePoint, offset_x, -8)
+      socketFrame:SetPoint(point, relativeTo, relativePoint, offsetX, -8)
     else
-      socketFrame:SetPoint(point, relativeTo, relativePoint, offset_x, 0)
+      socketFrame:SetPoint(point, relativeTo, relativePoint, offsetX, 0)
     end
   end
 
-  LCSFrame.Level:SetFont(font, fontSize, fontOutline)
-  LCSFrame.Level:SetTextColor(colorLevel.r, colorLevel.g, colorLevel.b, colorLevel.a)
-  LCSFrame.MaxLevel:SetFont(font, fontSize - 3, fontOutline)
-  LCSFrame.MaxLevel:SetTextColor(colorMaxLevel.r, colorMaxLevel.g, colorMaxLevel.b, colorMaxLevel.a)
-  LCSFrame.Enchant:SetFont(font, fontSize - 3, fontOutline)
-  LCSFrame.Tint:SetColorTexture(colorTexture.r, colorTexture.g, colorTexture.b, colorTexture.a)
+  slotOverlay.Level:SetFont(font, fontSize, fontOutline)
+  slotOverlay.Level:SetTextColor(colorLevel.r, colorLevel.g, colorLevel.b, colorLevel.a)
+  slotOverlay.MaxLevel:SetFont(font, fontSize - 3, fontOutline)
+  slotOverlay.MaxLevel:SetTextColor(colorMaxLevel.r, colorMaxLevel.g, colorMaxLevel.b, colorMaxLevel.a)
+  slotOverlay.Enchant:SetFont(font, fontSize - 3, fontOutline)
+  slotOverlay.Tint:SetColorTexture(colorTexture.r, colorTexture.g, colorTexture.b, colorTexture.a)
 
   if itemLevel ~= nil then
-    if itemLevel == MaxLevel or itemLevel >= maxUpgaradeLevel then
-      LCSFrame.Level:SetTextColor(colorMaxLevelUpgraded.r, colorMaxLevelUpgraded.g, colorMaxLevelUpgraded.b, colorMaxLevelUpgraded.a)
-      LCSFrame.MaxLevel:SetTextColor(colorMaxLevelUpgraded.r, colorMaxLevelUpgraded.g, colorMaxLevelUpgraded.b, colorMaxLevelUpgraded.a)
+    if itemLevel == maxLevel or itemLevel >= maxUpgradeLevel then
+      slotOverlay.Level:SetTextColor(colorMaxLevelUpgraded.r, colorMaxLevelUpgraded.g, colorMaxLevelUpgraded.b, colorMaxLevelUpgraded.a)
+      slotOverlay.MaxLevel:SetTextColor(colorMaxLevelUpgraded.r, colorMaxLevelUpgraded.g, colorMaxLevelUpgraded.b, colorMaxLevelUpgraded.a)
     end
   end
 
-  LCSFrame:Show()
+  slotOverlay:Show()
 end
 
+---@param unitId string
 local function UpdateAll(unitId)
   for slotId in pairs(Slots) do
     UpdateSlot(unitId, slotId)
   end
 end
 
+---@param self LCS_Events
+---@param event string
 local function OnEvent(self, event, ...)
-  if self[event] == nil then return end
+  if self[event] == nil then
+    return
+  end
   self[event](self, event, ...)
 end
 
-local f = CreateFrame("Frame")
+---@class LCS_Events : Frame
+local Events = CreateFrame("Frame")
+addon.Events = Events
 
-function f:PLAYER_ENTERING_WORLD()
+---@param self LCS_Events
+function Events:PLAYER_ENTERING_WORLD()
   UpdateAll("player")
 end
 
-function f:PLAYER_EQUIPMENT_CHANGED(_, slotId)
-  if slotId == nil then return end
+---@param self LCS_Events
+---@param slotId number|nil
+function Events:PLAYER_EQUIPMENT_CHANGED(_, slotId)
+  if slotId == nil then
+    return
+  end
   UpdateSlot("player", slotId)
 end
 
-function f:UNIT_INVENTORY_CHANGED(_, unitId)
-  if unitId == nil then return end
+---@param self LCS_Events
+---@param unitId string|nil
+function Events:UNIT_INVENTORY_CHANGED(_, unitId)
+  if unitId == nil then
+    return
+  end
   UpdateAll(unitId)
 end
 
-function f:INSPECT_READY()
+---@param self LCS_Events
+function Events:INSPECT_READY()
   UpdateAll("target")
 end
 
-function f:BAG_UPDATE_DELAYED()
+---@param self LCS_Events
+function Events:BAG_UPDATE_DELAYED()
   UpdateAll("player")
 end
 
-f:RegisterEvent("PLAYER_ENTERING_WORLD")
-f:RegisterEvent("PLAYER_EQUIPMENT_CHANGED")
-f:RegisterEvent("UNIT_INVENTORY_CHANGED")
-f:RegisterEvent("INSPECT_READY")
-f:RegisterEvent("BAG_UPDATE_DELAYED")
-f:SetScript("OnEvent", OnEvent)
+Events:RegisterEvent("PLAYER_ENTERING_WORLD")
+Events:RegisterEvent("PLAYER_EQUIPMENT_CHANGED")
+Events:RegisterEvent("UNIT_INVENTORY_CHANGED")
+Events:RegisterEvent("INSPECT_READY")
+Events:RegisterEvent("BAG_UPDATE_DELAYED")
+Events:SetScript("OnEvent", OnEvent)